float

+++ utils/casts/float.go
@@ -9,8 +9,8 @@ func Float32ToStr(f float32, prec int) string {
 	return strconv.FormatFloat(float64(f), 'f', prec, 64)
 }
 
-func Float64ToStr(f float32, prec int) string {
-	return strconv.FormatFloat(float64(f), 'f', prec, 64)
+func Float64ToStr(f float64, prec int) string {
+	return strconv.FormatFloat(f, 'f', prec, 64)
 }
 
 func FloatToInt(f float64) int {
